[Fix] #3 - 코드 간결화
