+++ MarketKurly-Clone/MarketKurly-Clone/Presentation/Home/ViewController/HomeViewController.swift
@@ -16,16 +16,12 @@ final class HomeViewController: UIViewController {
     // MARK: - UI Components
     
     private let navigationView = CustomNavigationBar()
-    private let menuCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        return collectionView
-    }()
+    private let menuCollectionView = MenuView()
     
     // MARK: - Properties
     
     private var menuTitleModel: [HomeMenuModel] = HomeMenuModel.homeMenuData()
+    private var menuCollectionViewInset: Int = 0
     
     // MARK: - Initializer
     
@@ -35,8 +31,6 @@ final class HomeViewController: UIViewController {
         super.viewDidLoad()
         setUI()
         setLayout()
-        setDelegate()
-        setRegister()
     }
 }
 
@@ -47,12 +41,7 @@ extension HomeViewController {
     private func setUI() {
         
         view.backgroundColor = Color.white
-        
-        menuCollectionView.do {
-            $0.isScrollEnabled = false
-            $0.showsHorizontalScrollIndicator = false
-            $0.backgroundColor = .clear
-        }
+    
     }
     
     // MARK: - Layout Helper
@@ -76,68 +65,5 @@ extension HomeViewController {
     
     // MARK: - Methods
     
-    private func setDelegate() {
-        menuCollectionView.dataSource = self
-        menuCollectionView.delegate = self
-    }
-    
-    private func setRegister() {
-        menuCollectionView.registerCell(MenuCollectionViewCell.self)
-    }
-    
-    func labelWidthSize(index: Int) -> Int {
-        let size = menuTitleModel[index].title.size(
-            withAttributes: [NSAttributedString.Key.font : UIFont.systemFont(ofSize: 15, weight: .regular)]).width
-        return Int(size)
-    }
-    
-    private func cellUnderLineSetting(cell: MenuCollectionViewCell?, indexPath: IndexPath, selected: Bool) {
-        cell?.isSelected = selected
-        cell?.setUnderLineWidth(size: labelWidthSize(index: indexPath.row))
-    }
-    
     // MARK: - @objc Methods
-}
-
-extension HomeViewController: UICollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return menuTitleModel.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueCell(type: MenuCollectionViewCell.self, indexPath: indexPath)
-        cell.setDataBind(model: menuTitleModel[indexPath.row])
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        let inset = SizeLiterals.Screen.screenWidth * 16 / 375
-        print(inset)
-        return UIEdgeInsets(top: 0, left: inset, bottom: 0, right: inset)
-    }
-}
-
-extension HomeViewController: UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let width = SizeLiterals.Screen.screenWidth * CGFloat(labelWidthSize(index: indexPath.row) + 20) / 375
-        print(width)
-        let height = CGFloat(44)
-        return CGSize(width: width, height: height)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return 0
-    }
-    
-    func  collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let cell = collectionView.cellForItem(at: indexPath) as? MenuCollectionViewCell
-        cellUnderLineSetting(cell: cell, indexPath: indexPath, selected: true)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
-        let cell = collectionView.cellForItem(at: indexPath) as? MenuCollectionViewCell
-        cellUnderLineSetting(cell: cell, indexPath: indexPath, selected: false)
-    }
 }
